python/django/requirements.txt
Django==4.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.13.0
requests==2.28.1
six==1.16.0
wheel==0.37.1
whitenoise==6.2.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1




